+++ package.json
@@ -26,9 +26,16 @@
     "@types/supertest": "^2.0.12",
     "@typescript-eslint/eslint-plugin": "^5.41.0",
     "@typescript-eslint/parser": "^5.60.1",
+    "@wdio/cli": "^8.12.2",
+    "@wdio/junit-reporter": "^8.11.0",
+    "@wdio/local-runner": "^8.12.1",
+    "@wdio/mocha-framework": "^8.12.1",
+    "@wdio/spec-reporter": "^8.12.2",
+    "@wdio/types": "^8.10.4",
     "aws-lambda": "^1.0.7",
     "aws-sdk-client-mock": "^2.0.0",
     "babel-jest": "^29.0.3",
+    "chromedriver": "^114.0.2",
     "copy-webpack-plugin": "^11.0.0",
     "csvtojson": "^2.0.10",
     "delay": "^5.0.0",
@@ -62,19 +69,12 @@
     "ts-loader": "^9.3.1",
     "ts-node": "^10.9.1",
     "typescript": "^4.9.5",
+    "wdio-chromedriver-service": "^8.1.1",
+    "wdio-typescript-service": "^0.0.3",
+    "webdriverio": "^8.11.2",
     "webpack": "^5.76.2",
     "webpack-cli": "^4.10.0",
-    "yaml": "^2.1.3",
-    "wdio-chromedriver-service": "^8.1.1",
-    "webdriverio": "^8.11.2",
-    "wdio-typescript-service": "^0.0.3",
-    "@wdio/types": "^8.10.4",
-    "@wdio/cli": "^8.12.2",
-    "@wdio/junit-reporter": "^8.11.0",
-    "@wdio/local-runner": "^8.12.1",
-    "@wdio/mocha-framework": "^8.12.1",
-    "@wdio/spec-reporter": "^8.12.2",
-    "chromedriver": "^114.0.2"
+    "yaml": "^2.1.3"
   },
   "scripts": {
     "husky:install": "husky install",
@@ -82,6 +82,7 @@
     "test:ephemeralEnvs": "jest --config jest.integration.config.ts --testPathIgnorePatterns email-storage-tests --maxWorkers=4 $JEST_ARGS; RET=$? ; npm run test:cleanup:reports ; exit $RET",
     "test:integration": "jest --config jest.integration.config.ts --maxWorkers=4 $JEST_ARGS; RET=$? ; npm run test:cleanup:reports ; exit $RET",
     "test:end2end": "jest --config jest.e2e.config.ts",
+    "test:ui-setup": "jest --config jest.ui.config.ts",
     "test:ui": "wdio run ./wdio.conf.ts",
     "test:cleanup:reports": "cd reports/jest-html-reports && ../../integration_tests/cleanup-jest-html-reports.sh 3",
     "test:integration:local": "jest --setupFiles=./integration_tests/setEnvVars.ts --testPathIgnorePatterns athena-query-tests s3-vat-tests --config jest.integration.config.ts",
